+++ application/src/App.css
@@ -6,7 +6,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: center;
   color: black;
 }
